feat: make order of fields in generated schema deterministic
Using HashMap in RootSchema makes the generated schema non-deterministic.
This can be painful when diffing against the previous version, for example when the generated schema is commited version control system.

This commit fixes that, by using BTreeMap for the RootSchema, which makes output fields always ordered lexicographically by key and thus deterministic.